+++ frontend/package.json
@@ -9,11 +9,16 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
+    "asyncawait": "^3.0.0",
     "bootstrap": "^4.4.1",
+    "cors": "^2.8.5",
     "emotion-theming": "^10.0.27",
+    "js-cookie": "^2.2.1",
+    "ky": "^0.19.1",
     "node-sass": "^4.14.0",
     "react": "^16.13.1",
     "react-bootstrap": "^1.0.1",
+    "react-cookie": "^4.0.3",
     "react-dom": "^16.13.1",
     "react-icons": "^3.10.0",
     "react-router-dom": "^5.1.2",
