+++ frontend/src/App.js
@@ -1,16 +1,17 @@
 import React, {Component} from 'react';
 import './App.css';
-import Login from './components/Login'
+import { BrowserRouter } from "react-router-dom";
+import AppRouter from "./AppRouter";
 
-class App extends Component {
-  render(){
-    return (
-      <div>
-        <Login />
+
+function App() {
+  return (
+    <div className="App">
+        <BrowserRouter>
+          <AppRouter />
+        </BrowserRouter>
       </div>
-    );
-  }
+  );
 }
-
 
 export default App;
