--- frontend/src/components/Login.css
@@ -1,75 +1,0 @@
-@media screen and (min-width:1700px){
-    .bkgd{
-        background-image:url('../images/backgroundv2.png');
-        background-size:auto;
-        height: 100vh;
-    }
-    h1.header{
-        text-align: center;
-        padding-top: 10vh;
-        font-size: 80px;
-        font-family: 'Raleway Light';
-        src:url('../fonts/Raleway-Light.ttf');
-        color: white;
-        text-shadow: 2px 2px 4px #000000;
-      }
-
-    .login-form{
-        width: 100%;
-        max-width:400px;
-        padding-top: 8vh;
-        max-height: 100vh;
-        margin: auto;
-        margin-bottom: 0;
-        position: sticky;
-    }
-
-    .box{
-        width: 100%;
-        padding:20px;
-        border-style:none;
-        border-radius: 20px;
-        box-shadow: 0 50px 50px 0 rgba(0, 0, 0, 0.2), 0 10px 30px 0 rgba(0, 0, 0, 0.19);
-        margin:auto;
-    }
-    h1.signIn{
-        text-align: center;
-        color: white;
-        font-family: 'Raleway Light';
-        src:url('../fonts/Raleway-Light.ttf');
-    }
-
-}
-
-@media screen and (max-width:800px){
-    .login-form{
-        width: 100%;
-        max-width:200px;
-        padding-top: 4vh;
-        height: 50vh;
-        margin: auto;
-        margin-bottom: 0;
-        position: sticky;
-    }
-
-    .box{
-        width: 100%;
-        padding:20px;
-        border-style:none;
-        /*background-color: #737373;*/
-        border-radius: 20px;
-        box-shadow: 0 50px 50px 0 rgba(0, 0, 0, 0.2), 0 10px 30px 0 rgba(0, 0, 0, 0.19);
-        margin:auto;
-    }
-    h1.signIn{
-        text-align: center;
-        color: white;
-        font-size: 20px;
-    }
-}
-
-a:link {
-    color: whitesmoke;
-    text-decoration: none;
-}
-
