--- frontend/src/components/Login.js
@@ -1,72 +1,0 @@
-import React, {Component} from 'react';
-import { Button, Form, FormGroup, Input, Container} from 'reactstrap';
-import './Login.css';
-import { Link } from 'react-router';
-
-
-class Login extends Component{
-  constructor(props){
-    super(props)
-    this.handleSubmit = this.handleSubmit.bind(this);
-    this.state={
-      email: "",
-      password: ""
-    }
-  }
-
-  handleSubmit = e => {
-    e.preventDefault()
-    fetch('/login', {
-      method: 'post',
-      headers: {
-          'Content-Type':'application/json',
-          'Accept': 'application/json'},
-      body: JSON.stringify({
-       "email": this.state.email,
-       "password": this.state.password
-      })
-    })
-    .then(response => response.json())
-    .then(data => this.setState({postId:data.id}));
-  }
-
-
-  render(){
-
-    return(
-      <Container className="bkgd" fluid>
-      <h1  className="header" >Photographer's portfolio</h1>
-        <Form className="login-form" onSubmit={this.handleSubmit}>
-        <div className="box">
-          <h1 className="signIn">Sign In</h1>
-        <FormGroup style={{paddingBottom:'5%', paddingTop:'5%'}}>
-          <Input value={this.state.email} onChange={(ev)=>this.setState({email:ev.target.value})} type="email" placeholder="Email"/>
-        </FormGroup>
-        <FormGroup style={{paddingBottom:'5%'}}>
-          <Input value={this.state.password} onChange={(ev)=>this.setState({password:ev.target.value})} type="password" placeholder="Password"/>
-        </FormGroup>
-        <Button type="submit" className="btn-lg btn-dark btn-block text-center" style={{textDecoration: 'none'}}>
-          <Link to="/mainPage" style={{textDecoration: 'none', color:'white'}}>
-          Sign in
-          </Link>
-        </Button>
-        <div className="text-center" style={{paddingTop:'8%'}}>
-        <Button to="/register" className="btn btn-dark" role="button">
-          <Link to="/mainPage" style={{textDecoration: 'none', color:'white'}}>
-          Sign Up
-          </Link>
-        </Button>
-        <span className="p-2"></span>
-        <a href="/forgot-password" className="btn btn-dark" role="button">Forgot Password</a>
-        </div>
-        <div className="text-center" style={{paddingTop:'5%', color:"white"}}>
-        By clicking “Sign up”, you agree to our <a href="#">Terms of Service</a> and <a href="#">Privacy Statement</a>.
-        </div>
-        </div>
-      </Form>
-      </Container>
-    );
-  }
-}
-
-export default Login;
