--- frontend/src/components/Register.scss
@@ -1,88 +1,0 @@
-@media screen and (min-width:1700px){
-  .bkgd{
-    background-image:url('../images/backgroundv2.png');
-    background-size:auto;
-    height: 100vh;
-    }
-
-    h1.header{
-        text-align: center;
-        padding-top: 10vh;
-        font-size: 80px;
-        font-family: 'Raleway Light';
-        src:url('../fonts/Raleway-Light.ttf');
-        color: white;
-        text-shadow: 2px 2px 4px #000000;
-    }
-    .register-form{
-        width: 100%;
-        max-width:800px;
-        padding-top: 5vh;
-        max-height: 100vh;
-        margin: auto;
-        margin-bottom: 0;
-        position: sticky;
-    }
-
-    .box{
-        border-radius: 20px;
-        box-shadow: 0 50px 50px 0 rgba(0, 0, 0, 0.2), 0 10px 30px 0 rgba(0, 0, 0, 0.19);
-        margin:auto;
-        padding:10px;
-    }
-    h1.signIn{
-        text-align: center;
-        color: white;
-        font-family: 'Raleway Light';
-        src:url('../fonts/Raleway-Light.ttf');
-
-    }
-
-   .vertical{
-    border-right: 1px solid white;
-    height: 620px;
-    position: absolute;
-
-   }
-   .vertical-divider {
-    position: absolute;
-    display: table;
-    text-align: center;
-    height: 100%;
-    width: 100%;
-
-    .split-layout__divider {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-
-        @media screen and (min-width: 30em) {
-         flex-direction: column;
-        }
-      }
-
-      .split-layout__label {
-        padding: 1em;
-        color: white;
-
-        font-size: 20px;
-        font-weight: bold;
-      }
-
-      .split-layout__rule {
-        flex: 1;
-        border-style: solid;
-        border-color: rgba(255,255,255,.3);
-        border-width: 1px 0 0 0;
-
-        @media screen and (min-width: 30em) {
-         border-width: 0 1px 0 0;
-        }
-      }
-    }
-
-    .pmd{
-        font-size: 20px;
-        color: white;
-    }
-}
